requirements.txt
pytelegrambotapi
python-dotenv
html2text